requirements.txt
# Langchain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable management
python-dotenv

# Machine learning utilities
numpy
scikit-learn


#ADD BELOW
#groq
#pip install torch transformers accelerate
#pip install -U sentence-transformers
#installed Pillow-11.3.0 sentence-transformers-5.1.1  